chore(web): code split into module files
