+++ src/commands/web/internal/spec/testFiles/script1.js
