+++ src/commands/web/web.ts
@@ -1,32 +1,22 @@
 import { chunk, getAbsolutePath } from '../../utils/utils'
 import {
   readFile,
-  copy,
-  base64EncodeFile,
   fileExists,
   folderExists,
   createFolder,
   createFile,
   deleteFolder,
-  listFilesInFolder,
   ServerType,
-  Target,
-  asyncForEach,
-  listSubFoldersInFolder
+  Target
 } from '@sasjs/utils'
 import { getStreamConfig } from '../../utils/config'
 import path from 'path'
-import jsdom, { JSDOM } from 'jsdom'
+import jsdom from 'jsdom'
 import { sasjsout } from './internal/sasjsout'
 import { adjustIframeScript } from './internal/adjustIframeScript'
-import { encode } from 'js-base64'
 import { StreamConfig } from '@sasjs/utils/types/config'
-import uniqBy from 'lodash.uniqby'
-
-const permittedServerTypes = {
-  SAS9: 'SAS9',
-  SASVIYA: 'SASVIYA'
-}
+import { createAssetServices } from './internal/createAssetServices'
+import { updateAllTags } from './internal/updateAllTags'
 
 const exampleStreamConfig: StreamConfig = {
   streamWeb: true,
@@ -41,6 +31,58 @@ export async function createWebAppServices(target: Target) {
 
   const streamConfig = await getStreamConfig(target)
 
+  const { webSourcePathFull, indexHtmlPath } = await validateStreamConfig(
+    streamConfig
+  )
+
+  process.logger?.info(
+    `Compiling web app services for target ${target.name}...`
+  )
+  await createBuildDestinationFolder()
+
+  const destinationPath = path.join(
+    buildDestinationServicesFolder,
+    streamConfig.streamWebFolder
+  )
+  await createTargetDestinationFolder(destinationPath)
+
+  const assetPathMap = await createAssetServices(
+    target,
+    destinationPath,
+    streamConfig
+  ).then((map) => {
+    const indexHtmlIndex = map.findIndex(
+      (assetPath) => assetPath.source === 'index.html'
+    )
+    map.splice(indexHtmlIndex, 1)
+    return map
+  })
+
+  const indexHtml = await readFile(indexHtmlPath).then(
+    (content) => new jsdom.JSDOM(content)
+  )
+
+  await updateAllTags(indexHtml, {
+    webSourcePathFull,
+    destinationPath,
+    serverType: target.serverType,
+    assetPathMap
+  })
+
+  if (target.serverType === ServerType.SasViya) {
+    indexHtml.window.document.body.innerHTML += adjustIframeScript
+    await createClickMeFile(
+      indexHtml.serialize(),
+      streamConfig.streamServiceName as string
+    )
+  } else
+    await createClickMeService(
+      indexHtml.serialize(),
+      streamConfig.streamServiceName as string
+    )
+}
+
+const validateStreamConfig = async (streamConfig: StreamConfig) => {
   if (!streamConfig) {
     throw new Error(
       `Invalid stream config: Please specify the \`streamConfig\` in your target in the following format: \n ${JSON.stringify(
@@ -72,346 +114,26 @@ export async function createWebAppServices(target: Target) {
     )
   }
 
-  process.logger?.info(
-    `Compiling web app services for target ${target.name}...`
+  const webSourcePathFull = getAbsolutePath(
+    streamConfig.webSourcePath,
+    process.projectDir
   )
-  await createBuildDestinationFolder()
 
-  const destinationPath = path.join(
-    buildDestinationServicesFolder,
-    streamWebFolder
-  )
-  await createTargetDestinationFolder(destinationPath)
-
-  const webSourcePathFull = getAbsolutePath(webSourcePath, process.projectDir)
-
-  if (!(await folderExists(webSourcePathFull)))
-    process.logger?.warn(
+  if (!(await folderExists(webSourcePathFull))) {
+    throw new Error(
       `webSourcePath: '${webSourcePathFull}' present in 'streamConfig' doesn't exist.`
     )
+  }
 
-  const assetPathMap = await createAssetServices(
-    target,
-    destinationPath,
-    streamConfig
-  )
   const indexHtmlPath = path.join(webSourcePathFull, 'index.html')
 
-  if (await fileExists(indexHtmlPath)) {
-    const indexHtmlIndex = assetPathMap.findIndex(
-      (assetPath) => assetPath.source === 'index.html'
-    )
-    assetPathMap.splice(indexHtmlIndex, 1)
-
-    const indexHtml = await readFile(indexHtmlPath).then(
-      (content) => new jsdom.JSDOM(content)
-    )
-
-    const assetsNotFound: Error[] = []
-
-    const scriptTags = getScriptTags(indexHtml)
-    await asyncForEach(scriptTags, async (tag) => {
-      await updateTagSource(
-        tag,
-        webSourcePath,
-        destinationPath,
-        target,
-        assetPathMap
-      ).catch((e) => assetsNotFound.push(e))
-    })
-
-    const linkTags = getLinkTags(indexHtml)
-    await asyncForEach(linkTags, async (linkTag) => {
-      await updateLinkHref(linkTag, assetPathMap).catch((e) =>
-        assetsNotFound.push(e)
-      )
-    })
-
-    const faviconTags = getFaviconTags(indexHtml)
-    await asyncForEach(faviconTags, async (faviconTag) => {
-      await updateLinkHref(faviconTag, assetPathMap).catch((e) =>
-        assetsNotFound.push(e)
-      )
-    })
-
-    if (assetsNotFound.length) {
-      const notFoundErrors = assetsNotFound
-        .map((e) => `- ${e.message}\n`)
-        .join('')
-      throw new Error(`Error(s) while preparing stream app:\n${notFoundErrors}`)
-    }
-
-    if (target.serverType === ServerType.SasViya) {
-      indexHtml.window.document.body.innerHTML += adjustIframeScript
-      await createClickMeFile(
-        indexHtml.serialize(),
-        streamConfig.streamServiceName as string
-      )
-    } else
-      await createClickMeService(
-        indexHtml.serialize(),
-        streamConfig.streamServiceName as string
-      )
-  }
-}
-
-async function createAssetServices(
-  target: Target,
-  destinationPath: string,
-  streamConfig: StreamConfig
-): Promise<{ source: string; target: string }[]> {
-  const { webSourcePath, streamWebFolder, assetPaths } = streamConfig
-  const assetPathMap: { source: string; target: string }[] = []
-
-  await asyncForEach(
-    [path.join(process.projectDir, webSourcePath), ...(assetPaths ?? [])],
-    async (assetPath) => {
-      const fullAssetPath = path.isAbsolute(assetPath)
-        ? assetPath
-        : path.isAbsolute(webSourcePath)
-        ? path.join(webSourcePath, assetPath)
-        : path.join(process.projectDir, webSourcePath, assetPath)
-      const assetPathExists = await folderExists(fullAssetPath)
-
-      if (!assetPathExists) {
-        process.logger?.warn(
-          `Assets path '${fullAssetPath}' present in 'streamConfig' doesn't exist.`
-        )
-        return
-      }
-      const assetPathMapNested = await createAssetsServicesNested(
-        target,
-        streamWebFolder,
-        fullAssetPath,
-        destinationPath
-      )
-      assetPathMap.push(...assetPathMapNested)
-    }
-  )
-  return uniqBy(assetPathMap, 'source')
-}
-
-async function createAssetsServicesNested(
-  target: Target,
-  streamWebFolder: string,
-  fullAssetPath: string,
-  destinationPath: string
-) {
-  const assetPathMap: { source: string; target: string }[] = []
-  const filePaths = await listFilesInFolder(fullAssetPath)
-  await asyncForEach(filePaths, async (filePath) => {
-    const fullFileName = path.basename(filePath)
-    const fileName = fullFileName.substring(0, fullFileName.lastIndexOf('.'))
-    const fileExtension = path
-      .basename(filePath)
-      .substring(fullFileName.lastIndexOf('.') + 1, fullFileName.length)
-    if (fileName && fileExtension) {
-      const sourcePath = path.join(fullAssetPath, filePath)
-      if (target.serverType === ServerType.SasViya) {
-        await copy(sourcePath, path.join(destinationPath, fullFileName))
-        const assetServiceUrl = getAssetPath(
-          target.appLoc,
-          target.serverType,
-          streamWebFolder,
-          fullFileName
-        )
-        assetPathMap.push({
-          source: fullFileName,
-          target: assetServiceUrl
-        })
-      } else {
-        const fileName = await generateAssetService(
-          sourcePath,
-          filePath,
-          destinationPath,
-          target.serverType
-        )
-        const assetServiceUrl = getAssetPath(
-          target.appLoc,
-          target.serverType,
-          streamWebFolder,
-          fileName.replace('.sas', '')
-        )
-        assetPathMap.push({
-          source: fullFileName,
-          target: assetServiceUrl
-        })
-      }
-    }
-  })
-
-  const folderNames = await listSubFoldersInFolder(fullAssetPath)
-  await asyncForEach(folderNames, async (folderName) => {
-    const destinationPathNested = path.join(destinationPath, folderName)
-    const fullAssetPathNested = path.join(fullAssetPath, folderName)
-
-    await createFolder(destinationPathNested)
-
-    const assetFolderMap: { source: string; target: string }[] =
-      await createAssetsServicesNested(
-        target,
-        `${streamWebFolder}/${folderName}`,
-        fullAssetPathNested,
-        destinationPathNested
-      )
-    assetFolderMap.forEach((entry) => {
-      assetPathMap.push({
-        source: `${folderName}/${entry.source}`,
-        target: entry.target
-      })
-    })
-  })
-  return assetPathMap
-}
-
-async function generateAssetService(
-  sourcePath: string,
-  filePath: string,
-  destinationPath: string,
-  serverType: ServerType
-) {
-  const fileExtension = path.extname(filePath)
-  const fileType = fileExtension.replace('.', '').toUpperCase()
-  const fileName = path
-    .basename(filePath)
-    .replace(new RegExp(fileExtension + '$'), fileExtension.replace('.', '-'))
-  const base64string = await base64EncodeFile(sourcePath)
-
-  const serviceContent = await getWebServiceContentSAS9(
-    base64string,
-    fileType,
-    serverType
-  )
-
-  await createFile(
-    path.join(destinationPath, `${fileName}.sas`),
-    serviceContent
-  )
-
-  return `${fileName}.sas`
-}
-
-export async function updateTagSource(
-  tag: HTMLLinkElement,
-  webAppSourcePath: string,
-  destinationPath: string,
-  target: Target,
-  assetPathMap: { source: string; target: string }[]
-) {
-  const scriptPath = tag.getAttribute('src')
-  const isUrl =
-    scriptPath && (scriptPath.startsWith('http') || scriptPath.startsWith('//'))
-
-  if (scriptPath) {
-    if (!isUrl) {
-      let content = await readFile(
-        path.join(process.projectDir, webAppSourcePath, scriptPath)
-      ).catch((_) => {
-        throw new Error(`Unable to find file: ${scriptPath}`)
-      })
-
-      assetPathMap.forEach((pathEntry) => {
-        content = content.replace(
-          new RegExp(pathEntry.source, 'g'),
-          pathEntry.target
-        )
-      })
-
-      if (target.serverType === ServerType.SasViya) {
-        await createFile(path.join(destinationPath, scriptPath), content)
-      } else {
-        const serviceContent = await getWebServiceContentSAS9(
-          encode(content),
-          'JS',
-          target.serverType
-        )
-        await createFile(
-          path.join(
-            destinationPath,
-            `${scriptPath.replace(/\.js$/, '-js')}.sas`
-          ),
-          serviceContent
-        )
-      }
-
-      tag.setAttribute(
-        'src',
-        assetPathMap.find(
-          (entry) =>
-            entry.source === scriptPath || `./${entry.source}` === scriptPath
-        )!.target
-      )
-    }
-  }
-}
-
-async function updateLinkHref(
-  linkTag: HTMLLinkElement,
-  assetPathMap: { source: string; target: string }[]
-) {
-  const linkSourcePath = linkTag.getAttribute('href') || ''
-  const isUrl =
-    linkSourcePath.startsWith('http') || linkSourcePath.startsWith('//')
-  if (!isUrl) {
-    const assetPath = assetPathMap.find(
-      (entry) =>
-        entry.source === linkSourcePath ||
-        `./${entry.source}` === linkSourcePath
-    )
-
-    if (!assetPath?.target) {
-      throw new Error(`Unable to find file: ${linkSourcePath}`)
-    }
-
-    linkTag.setAttribute('href', assetPath.target)
-  }
-}
-
-function getAssetPath(
-  appLoc: string,
-  serverType: ServerType,
-  streamWebFolder: string,
-  fileName: string
-) {
-  if (!permittedServerTypes.hasOwnProperty(serverType.toUpperCase())) {
+  if (!(await fileExists(indexHtmlPath))) {
     throw new Error(
-      'Unsupported server type. Supported types are SAS9 and SASVIYA'
+      `'index.html' doesn't exist at webSourcePath: '${webSourcePathFull}' present in 'streamConfig'`
     )
   }
-  const storedProcessPath =
-    // the appLoc is inserted dynamically by SAS
-    // using three forward slashes as a marker
-    // for SAS 9 fileName is a program, with replacement in sasjsout.ts
-    // for Viya, fileName is a FILE, with replacement in build.sas only
-    serverType === ServerType.SasViya
-      ? `/SASJobExecution?_FILE=${appLoc}/services/${streamWebFolder}`
-      : `/SASStoredProcess/?_PROGRAM=///${streamWebFolder}`
-  return `${storedProcessPath}/${fileName}`
-}
 
-function getScriptTags(parsedHtml: JSDOM) {
-  return Array.from(parsedHtml.window.document.querySelectorAll('script'))
-}
-
-function getLinkTags(parsedHtml: JSDOM) {
-  const linkTags = Array.from(
-    parsedHtml.window.document.querySelectorAll('link')
-  ).filter((s) => s.getAttribute('rel') === 'stylesheet')
-
-  return linkTags
-}
-
-function getFaviconTags(parsedHtml: JSDOM) {
-  const linkTags = Array.from(
-    parsedHtml.window.document.querySelectorAll('link')
-  ).filter(
-    (s) =>
-      s.getAttribute('rel') &&
-      s.getAttribute('rel') &&
-      s.getAttribute('rel')!.includes('icon')
-  )
-
-  return linkTags
+  return { webSourcePathFull, indexHtmlPath }
 }
 
 async function createBuildDestinationFolder() {
@@ -428,55 +150,6 @@ async function createTargetDestinationFolder(destinationPath: string) {
     await deleteFolder(destinationPath)
   }
   await createFolder(destinationPath)
-}
-
-async function getWebServiceContentSAS9(
-  contentBase64: string,
-  type = 'JS',
-  serverType: ServerType
-) {
-  let lines = [contentBase64]
-
-  // Encode to base64 *.js and *.css files if target server type is SAS 9.
-  const typesToEncode: { [key: string]: string } = {
-    JS: 'JS64',
-    CSS: 'CSS64'
-  }
-
-  let serviceContent = `${sasjsout}\nfilename sasjs temp lrecl=99999999;
-data _null_;
-file sasjs;
-`
-
-  lines.forEach((line) => {
-    const chunkedLines = chunk(line)
-    if (chunkedLines.length === 1) {
-      serviceContent += `put '${chunkedLines[0].split("'").join("''")}';\n`
-    } else {
-      let combinedLines = ''
-      chunkedLines.forEach((chunkedLine, index) => {
-        let text = `put '${chunkedLine.split("'").join("''")}'`
-        if (index !== chunkedLines.length - 1) {
-          text += '@;\n'
-        } else {
-          text += ';\n'
-        }
-        combinedLines += text
-      })
-      serviceContent += combinedLines
-    }
-  })
-
-  if (
-    serverType === permittedServerTypes.SAS9 &&
-    typesToEncode.hasOwnProperty(type)
-  ) {
-    serviceContent += `\nrun;\n%sasjsout(${typesToEncode[type]})`
-  } else {
-    serviceContent += `\nrun;\n%sasjsout(${type})`
-  }
-
-  return serviceContent
 }
 
 async function createClickMeService(
